Candy Catch Game
I used the given assets to create a basic layout of the game. Then I added the Input System and connected it to the Monster image to become the player using the correct code. The next steps are not functioning well for me and I ran out of time so I just polished the game to end it off.